requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
anyio==4.4.0
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.6.2
exceptiongroup==1.2.1
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
multidict==6.0.5
playwright==1.44.0
pyee==11.1.0
python-telegram-bot==21.3
sniffio==1.3.1
soupsieve==2.5
typing_extensions==4.12.2
yarl==1.9.4
